Organização do Projeto + Correções do Back-End
- Organização dos módulos, pacotes e arquitetura do projeto

- Correções do back-end
