+++ modules/config.py
@@ -1,0 +1,5 @@
+import os
+import binascii
+
+SECRET_KEY = binascii.hexlify(os.urandom(24))
+API_KEY = '7IEKBZD1IV0BSJQN'
